requirements.txt
datasets==1.14.0
lib==3.0.0
numpy==1.20.1
opencv_python==4.5.3.56
Pillow==8.4.0
PyYAML==6.0
scikit_image==0.18.1
scipy==1.6.2
skimage==0.0
tensorboardX==2.4